fixed undefined behaviors in test code

+++ src/binary64/tanpi/check_special.c
@@ -76,7 +76,7 @@ double rand_arg(double s){
   int64_t r0,r1;
   r0 = rand () | (int64_t) rand () << 31;
   r1 = rand () | (int64_t) rand () << 31;
-  b64u64_u o = {.u = (((r0^(r1<<32))&(~(0x7ffull<<52)))|(0x3ffull<<52))};
+  b64u64_u o = {.u = (((r0^((uint64_t)r1<<32))&(~(0x7ffull<<52)))|(0x3ffull<<52))};
   double r = o.f-copysign(1,o.f);
   return r*s;
 }
